+++ lib/src/user_location_options.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/plugin_api.dart';
-import 'package:latlong/latlong.dart';
-import 'dart:async';
-import '../utils/getLocation.dart';
 
 class UserLocationOptions extends LayerOptions {
   BuildContext context;
   MapController mapController;
   List<Marker> markers;
-  UserLocationOptions({this.context, this.mapController, this.markers});
+  Widget markerWidget;
+  UserLocationOptions(
+      {this.context, this.mapController, this.markers, this.markerWidget});
 }