+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## 0.0.2
 
 * Changed author email id
+
+## 0.0.3
+
+* Added marker to show accuracy
+* Custom user marker widget
+* Removed unnecessary files